+++ src/main/java/com/jino/realbread/domain/user/repository/UserRepository.java
@@ -12,7 +12,7 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     Optional<UserEntity> findByOauthId(String oauthId);
 
-    boolean existsByUserId(Integer userId);
+    boolean existsByUserId(Long userId);
 
     UserEntity findByEmail(String email);
 }
